Hotkey logic changed
Hotkeys are now compiled functions to check whenever an event is a hotkey. Also, addHotkeys() now accepts [key, callback] pairs to be checked.
